feat(rdma): multi-NIC path awareness with load-balanced placement and rebalancing (#77)
Peers stay identified by their bootstrap TCP address, but every RDMA
connection (stripe) now carries its path — the (local NIC, remote NIC)
pair — and picks it independently:

- Local NIC by least-connections over live per-device counters
  (outbound + inbound, RAII-accurate via ConnCountGuard).
- Remote NIC by power-of-two-choices over the peer's advertised
  per-NIC load (new RdmaDeviceInfo.active_connections), avoiding
  client herding on stale snapshots.
- Device matching cannot verify routability: QP setup failures
  blacklist the NIC pair per peer for 30s and placement falls over
  to the next candidate.

A per-pool maintenance task (jittered rdma.maintenance_interval_ms)
fails connections on downed local ports, prunes dead stripes,
replenishes peers up to connections_per_peer, and migrates at most
one connection per tick onto a less-loaded pair (make-before-break
with a drain grace period; self-excluding scores plus
rebalance_threshold hysteresis prevent oscillation). Rate-limited
migration makes traffic return gradually to recovered NICs.

Explicit control and introspection:
- Context::with_rdma_path(RdmaPathSelector) pins requests to paths
  matching a device name or GID-embedded IP, creating pinned
  (rebalance-exempt) connections on demand.
- rdma.remote_device_filter complements the local device_filter.
- State::rdma_path_report() / Server::state() expose every live
  connection with its NIC pair, direction, health and per-device
  load; ConnectRequest.source_device gives servers the client-side
  NIC of inbound connections.

+++ ruapc-rdma/src/verbs/context.rs
@@ -60,7 +60,7 @@ impl Context {
     }
 
     /// Queries a GID for the specified port and index.
-    pub fn query_gid(&self, port_num: u8, gid_index: u16) -> Result<crate::ibv_gid> {
+    pub fn query_gid(&self, port_num: u8, gid_index: u8) -> Result<crate::ibv_gid> {
         let mut gid = crate::ibv_gid::default();
         let ret =
             unsafe { crate::ibv_query_gid(self.ptr, port_num as _, gid_index as _, &mut gid) };
@@ -75,7 +75,7 @@ impl Context {
     pub fn query_gid_type(
         &self,
         port_num: u8,
-        gid_index: u16,
+        gid_index: u8,
         ibdev_path: &Path,
         port_attr: &crate::ibv_port_attr,
     ) -> Result<GidType> {
